+++ packages/iov-jsonrpc/src/jsonrpcclient.ts
@@ -1,6 +1,6 @@
 import { Stream } from "xstream";
 
-import { toListPromise } from "@iov/stream";
+import { firstEvent } from "@iov/stream";
 
 import { isJsonRpcErrorResponse, JsonRpcRequest, JsonRpcResponse, JsonRpcSuccessResponse } from "./types";
 
@@ -18,10 +18,10 @@ export class JsonRpcClient {
 
   public async run(request: JsonRpcRequest): Promise<JsonRpcSuccessResponse> {
     const filteredStream = this.connection.responseStream.filter(r => r.id === request.id);
-    const pendingResponses = toListPromise(filteredStream, 1);
+    const pendingResponses = firstEvent(filteredStream);
     this.connection.sendRequest(request);
 
-    const response = (await pendingResponses)[0];
+    const response = await pendingResponses;
     if (isJsonRpcErrorResponse(response)) {
       const error = response.error;
       throw new Error(`JSON RPC error: code=${error.code}; message='${error.message}'`);