+++ packages/iov-stream/src/index.ts
@@ -1,5 +1,5 @@
 export { DefaultValueProducer, DefaultValueProducerCallsbacks } from "./defaultvalueproducer";
 export { concat } from "./concat";
-export { fromListPromise, toListPromise } from "./promise";
+export { firstEvent, fromListPromise, toListPromise } from "./promise";
 export * from "./reducer";
 export { ValueAndUpdates } from "./valueandupdates";